Armraise function added, non-functional (Decrease range in > and < checks)

+++ Finished_Files/Project/Rand_Func.ino
@@ -1,6 +1,6 @@
 #include "Rand_Func.h"
 
-int randomise(float num){
+int randomise(int num){
   return millis() % num;
 }
